big terrain rendering fixed - u16 indices were problem - mesh buffer generation now uses more mesh buffers

+++ irrlicht_ode/terrain.h
@@ -44,17 +44,13 @@ public:
 	// SMeshBuffer. This function chops it into pieces and generates a
 	// buffer from each one.
 
-	//void init(HeightMap &hm, f32 scale, colour_func cf, IVideoDriver *driver);
-
 	// Generate a SMeshBuffer which represents all the vertices and
 	// indices for values of y between y0 and y1, and add it to the
 	// mesh.
 
-	//void addstrip(const HeightMap &hm, colour_func cf, u16 y0, u16 y1, u32 bufNum);
-
-	void addstrip(pixmap *pixmapInstance, colour_func cf, u16 y0, u16 y1, u32 bufNum);
+	void addstrip(pixmap *pixmapInstance, colour_func cf, u32 y0, u32 y1, u32 bufNum);
 	void init(pixmap *pixmapInstance, f32 scale, colour_func cf, IVideoDriver *driver);
-	core::vector3df GetNormal(u16 x, u16 y, f32 s) const;
+	core::vector3df GetNormal(u32 x, u32 y, f32 s) const;
 
 	f32 Get(int x, int y) const
 	{
